Big update

+++ src/main/java/com/epam/hospital/controller/command/impl/Consultation/ConsultationApproveCommand.java
@@ -9,8 +9,8 @@ import com.epam.hospital.controller.request.RequestContext;
 import com.epam.hospital.model.treatment.Consultation;
 import com.epam.hospital.model.treatment.type.CommunicationType;
 import com.epam.hospital.model.treatment.type.ConsultationStatus;
-import com.epam.hospital.service.database.ConsultationService;
-import com.epam.hospital.service.database.impl.ConsultationServiceImpl;
+import com.epam.hospital.service.logic.ConsultationService;
+import com.epam.hospital.service.logic.impl.ConsultationServiceImpl;
 import com.epam.hospital.service.exception.ServiceException;
 import lombok.extern.slf4j.Slf4j;
 
@@ -20,7 +20,7 @@ public class ConsultationApproveCommand implements Command {
     private static final String CONSULTATION_PAGE_COMMAND = "MentalHospital?command=" + CommandName.CONSULTATION_PAGE;
     private static final String CONSULTATION_STATUS_APPROVED = "APPROVED";
 
-    private static final int DEFAULT_PRICE = 10;
+    private static final double DEFAULT_PRICE = 10;
 
     @Override
     public CommandResult execute(RequestContext requestContext) throws ServiceException {
