+++ src/main/java/com/epam/hospital/dao/builder/impl/ConsultationBuilder.java
@@ -19,7 +19,7 @@ public class ConsultationBuilder implements EntityBuilder<Consultation> {
         consultation.setDuration(resultSet.getInt(Column.CONSULTATION_DURATION));
         consultation.setDoctorId(resultSet.getInt(Column.CONSULTATION_DOCTOR_ID));
         consultation.setPatientId(resultSet.getInt(Column.CONSULTATION_PATIENT_ID));
-        consultation.setPrice(resultSet.getInt(Column.CONSULTATION_PRICE));
+        consultation.setPrice(resultSet.getDouble(Column.CONSULTATION_PRICE));
         consultation.setTreatmentCourseId(resultSet.getInt(Column.CONSULTATION_TREATMENT_COURSE_ID));
         return consultation;
     }
